feat: make io filler constructor
Adding this constructor allows us to remove the direct dependency
on the environschema package. The idmclient can offer out a
form.Filler using the environschema package, with the right std
io dependencies. Thus we only directly depend on std lib io package.

Once we remove macaroons in the distant future, we can just rip out
the idm client et al and we won't depend on this or the
environschema anymore.